+++ api-v2/app/com/foreignlanguagereader/api/domain/definition/entry/webster/common/WebsterMeta.scala
@@ -1,0 +1,35 @@
+package com.foreignlanguagereader.api.domain.definition.entry.webster.common
+
+import com.foreignlanguagereader.api.domain.definition.entry.webster.common.WebsterSource.WebsterSource
+import play.api.libs.functional.syntax._
+import play.api.libs.json.{JsPath, Json, Reads, Writes}
+
+case class WebsterMeta(id: String,
+                       uuid: String,
+                       sort: String,
+                       source: WebsterSource,
+                       section: String,
+                       stems: Seq[String],
+                       offensive: Boolean)
+object WebsterMeta {
+  implicit val writes: Writes[WebsterMeta] = Json.writes[WebsterMeta]
+  implicit val reads: Reads[WebsterMeta] = (
+    (JsPath \ "id").read[String] and
+      (JsPath \ "uuid").read[String] and
+      (JsPath \ "sort").read[String] and
+      (JsPath \ "src").read[WebsterSource] and
+      (JsPath \ "section").read[String] and
+      (JsPath \ "stems").read[Seq[String]](Reads.seq[String]) and
+      (JsPath \ "offensive").read[Boolean]
+  )(WebsterMeta.apply _)
+}
+
+object WebsterSource extends Enumeration {
+  type WebsterSource = Value
+
+  val LEARNERS: Value = Value("learners")
+  val SPANISH: Value = Value("spanish")
+
+  implicit val reads: Reads[WebsterSource] = Reads.enumNameReads(WebsterSource)
+  implicit val writes: Writes[WebsterSource] = Writes.enumNameWrites
+}
